prsv_project/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
scikit-image==0.24.0
scikit-learn==1.5.2
matplotlib==3.9.2
pandas==2.2.3
joblib==1.4.2

pytest==8.3.3
httpx==0.27.2